refactoring mapbox js code in a separate file for cleaningness
